+++ gamemodes/terrortown/gamemode/client/cl_equip.lua
@@ -1,6 +1,6 @@
 ---
 -- Traitor equipment menu
--- @section stop_manager
+-- @section shop
 
 local GetTranslation = LANG.GetTranslation
 local GetPTranslation = LANG.GetParamTranslation
@@ -51,8 +51,71 @@ local serverColsVar = GetConVar("ttt_bem_sv_cols")
 local serverRowsVar = GetConVar("ttt_bem_sv_rows")
 local serverSizeVar = GetConVar("ttt_bem_sv_size")
 
--- add favorites DB functions
-include("favorites_db.lua")
+---
+-- Some database functions of the shop
+
+---
+-- Creates the fav table if it not already exists
+-- @realm client
+local function CreateFavTable()
+	if not sql.TableExists("ttt_bem_fav") then
+		local query = "CREATE TABLE ttt_bem_fav (guid TEXT, role TEXT, weapon_id TEXT)"
+		sql.Query(query)
+	end
+end
+
+---
+-- Adds a @{WEAPON} or an @{ITEM} into the fav table
+-- @param number steamid steamid of the @{Player}
+-- @param number subrole subrole id
+-- @param string id the @{WEAPON} or @{ITEM} id
+-- @realm client
+local function AddFavorite(steamid, subrole, id)
+	local query = ("INSERT INTO ttt_bem_fav VALUES('" .. steamid .. "','" .. subrole .. "','" .. id .. "')")
+	sql.Query(query)
+end
+
+---
+-- Removes a @{WEAPON} or an @{ITEM} into the fav table
+-- @param number steamid steamid of the @{Player}
+-- @param number subrole subrole id
+-- @param string id the @{WEAPON} or @{ITEM} id
+-- @realm client
+local function RemoveFavorite(steamid, subrole, id)
+	local query = ("DELETE FROM ttt_bem_fav WHERE guid = '" .. steamid .. "' AND role = '" .. subrole .. "' AND weapon_id = '" .. id .. "'")
+	sql.Query(query)
+end
+
+---
+-- Get all favorites of a @{Player} based on the subrole
+-- @param number steamid steamid of the @{Player}
+-- @param number subrole subrole id
+-- @return table list of all favorites based on the subrole
+-- @realm client
+local function GetFavorites(steamid, subrole)
+	local query = ("SELECT weapon_id FROM ttt_bem_fav WHERE guid = '" .. steamid .. "' AND role = '" .. subrole .. "'")
+
+	return sql.Query(query)
+end
+
+---
+-- Looks for weapon id in favorites table (result of GetFavorites)
+-- @param table favorites favorites table
+-- @param number id id of the @{WEAPON} or @{ITEM}
+-- @return boolean
+-- @realm client
+local function IsFavorite(favorites, id)
+	for _, value in pairs(favorites) do
+		local dbid = value["weapon_id"]
+
+		if dbid == tostring(id) then
+			return true
+		end
+	end
+
+	return false
+end
+
 
 local color_bad = Color(244, 67, 54, 255)
 --local color_good = Color(76, 175, 80, 255)