+++ gamemodes/terrortown/gamemode/shared/sh_item_module.lua
@@ -1,8 +1,5 @@
 local oldITEM = ITEM
 
--- include modules
-require("items")
-
 -- load items
 local itemsPre = "terrortown/entities/items/"
 local itemsFiles = file.Find(itemsPre .. "*.lua", "LUA")