+++ lua/terrortown/lang/ru.lua
@@ -1412,3 +1412,6 @@ L.desc_event_kill_other_using = "{victim} ({vrole} / {vteam}) был убит {a
 --2021-06-15
 --L.header_shop_linker = "Settings"
 --L.label_shop_linker_set = "Shopsetting"
+
+--2021-08-18
+--L.xfer_team_indicator = "Team"