+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 
 ## Unreleased
 
+## [v0.9.0b](https://github.com/TTT-2/TTT2/tree/v0.9.0b) (2021-06-19)
+
 ### Added
 
 - All new roundend menu
@@ -16,6 +18,7 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
   - Enables the possibility to give Innocents access to a custom shop (`shopeditor`)
 - Karma now stores changes
   - Is shown in roundend menu
+- Added a new hook `TTT2ModifyLogicCheckRole` that can be used to modify the tested role for map related role checks
 - Added the ConVar `ttt2_random_shop_items` for the number of items in the randomshop
 - Added per-player voice control by hovering over the mute icon and scrolling
 
@@ -31,14 +34,20 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 - Fixed issue where roles with unknownTeam could see messages sent with the team chat key
 - Fixed the admin section label not being visible in the main menu
 - Fixed the auto resizing of the buttons based on the availability of a scrollbar not working
+- Fixed reopening submenus of the legacy addons in F1 menu not working
+- TTT: Fixed karma autokick evasion
+- TTT: Fixed karma being applied to weapon damage even though karma is disabled
 
 ### Changed
 
 - Microoptimization to improve code performance
-- Converted `roles` module into a library
+- Converted `roles`, `huds`, `hudelements`, `items` and `pon` modules into libraries
+- Moved `bind` library to the libraries folder
+- Moved favorites functions for equipment to the equipment shop and made them local functions
 - Code cleanup and removed silly negations
 - Extended some ttt2net functions
 - Changed `bees` win to `nones` win
+- By default all evil roles are now counted as traitor roles for map related checks
 - Changed the ConVar `ttt2_random_shops` to only disable the random shop (if set to `0`)
 - Shopeditor settings are now available in the F1 Menu
 - Moved the F1 menu generating system from a hook based system to a file based system
@@ -51,6 +60,8 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 - Shopeditor equipment is now available in F1 menu
 - Moved the role layering menu to the F1 menu (administration submenu)
   - removed the command `ttt2_edit_rolelayering`
+- moved the internal path of `lang/`, `vskin/` and `events/` (this doesn't change anything for addons)
+- Sort teammates first in credit transfer selection and add an indicator to them
 
 ### Removed
 
@@ -61,6 +72,7 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 - Adjusted `Player:HasRole()` and `Player:HasTeam()` to support simplified role and team checks (no parameter are supported anymore, use `Player:GetRole()` or `Player:GetTeam()` instead)
 - Moved global roleData to the `roles` library (e.g. `INNOCENT` to `roles.INNOCENT`). `INNOCENT`, `TRAITOR` etc. is not supported anymore. `ROLE_<ROLENAME>` is still supported and won't be changed.
 - Shopeditor function `ShopEditor.ReadItemData()` now only updates a number of key-parameters, must be given as UInt. Messages were changed accordingly (`TTT2SESaveItem`,`TTT2SyncDBItems`)
+- Equipment shop favorite functions are now local and not global anymore (`CreateFavTable`, `AddFavorite`, `RemoveFavorite`, `GetFavorites` & `IsFavorite`)
 
 
 ## [v0.8.2b](https://github.com/TTT-2/TTT2/tree/v0.8.2b) (2021-03-25)